added src,component folder and separate components
