+++ src/App.js
@@ -1,27 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-
-//Header
-const Header = () => {
-  return (
-    <div className="header">
-      <div className="logo-container">
-        <img
-          className="logo"
-          src="https://www.logodesign.net/logo/smoking-burger-with-lettuce-3624ld.png"
-        />
-      </div>
-      <div className="nav-item">
-        <ul>
-          <li>Home</li>
-          <li>About Us</li>
-          <li>Conatct Us</li>
-          <li>Cart</li>
-        </ul>
-      </div>
-    </div>
-  );
-};
+import Header from '../src/component/Header'
 
 const restaurantList = [
   {
@@ -1876,7 +1855,7 @@ const Body = () => {
          ))
        }
       </div>
-    </div>
+    </div>  
   );
 };
